Add ModelClassField and move UserModel to common schema module
The ModelClassField can be reused to declare the common property model_class field in pydantic models
