+++ lib/galaxy/managers/users.py
@@ -9,7 +9,6 @@ import time
 from datetime import datetime
 
 from markupsafe import escape
-from pydantic import BaseModel, Field
 from sqlalchemy import and_, desc, exc, func, true
 from sqlalchemy.orm.exc import NoResultFound
 
@@ -48,17 +47,6 @@ you may want to notify an administrator.
 If you're having trouble using the link when clicking it from email client, you
 can also copy and paste it into your browser.
 """
-
-
-class UserModel(BaseModel):
-    """User in a transaction context."""
-    id: int = Field(title='ID', description='User ID')
-    username: str = Field(title='Username', description='User username')
-    email: str = Field(title='Email', description='User email')
-    active: bool = Field(title='Active', description='User is active')
-    deleted: bool = Field(title='Deleted', description='User is deleted')
-    last_password_change: datetime = Field(title='Last password change', description='')
-    model_class: str = Field(title='Model class', description='Database model class (User)')
 
 
 class UserManager(base.ModelManager, deletable.PurgableManagerMixin):
